+++ CHANGELOG.md
@@ -107,6 +107,7 @@
 - Add content rule condition handlers @robgietema
 - Add content rule action handlers @robgietema
 - Add rss support @robgietema
+- Add feed content type @robgietema
 
 ### Bugfix